Add support for NODE_ENV
NODE_ENV is an environment variable made popular by the express web server framework. When a node application is run, it can check the value of the environment variable and do different things based on the value. NODE_ENV specifically is used (by convention) to state whether a particular environment is a production or a development environment.
NODE_ENV is added to the config files to activate/deactivate returning System error messages to API clients based on the execution mode.
The following code is used to access the environment variable:
const isDevelopment = process.env.NODE_ENV === 'development';

You can alternatively using express' app.get('env') function, but note that this is NOT RECOMMENDED as it defaults to "development", which may result in development code being accidentally run in a production environment. Conversely, set NODE_ENV for your environment, it will be undefined if you access it from process.env, there is no default.
 Changes to be committed:
	modified:   app.js
	modified:   connections/defaultConnect.js
	modified:   models/user.js
	modified:   package.json